refactor specs and support files
