+++ spec/action_presenter/base_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-require 'support/base'
+require 'support/dummy_presenter'
 
 
 describe ActionPresenter::Base do
