+++ spec/support/dummy_action_controller.rb
@@ -1,0 +1,17 @@
+class DummyActionController
+  def self.helper_method(name)
+    nil
+  end
+
+  def view_context
+    self
+  end
+end
+
+module ActionControllerHelper
+  extend ActiveSupport::Concern
+
+  included do
+    let(:helper) { DummyActionController.send(:include, ActionPresenter::ViewHelper).new }
+  end
+end
