+++ spec/support/string_presenter.rb
@@ -1,13 +1,3 @@
-class DummyActionController
-  def self.helper_method(name)
-    nil
-  end
-
-  def view_context
-    self
-  end
-end
-
 class StringPresenter < ActionPresenter::Base
   def foobar
     'foo'
